pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pystatkit"
version = "0.1.0"
description = "A reproducible, human-in-the-loop statistical analysis toolkit for behavioural and health sciences research."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Joash Ye" },
]
keywords = ["statistics", "research", "APA", "ANOVA", "reproducibility"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "pingouin>=0.5.4",
    "statsmodels>=0.14",
    "tableone>=0.9",
    "pyyaml>=6.0",
    "python-docx>=1.1",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
pystatkit = "pystatkit.cli:main"

[tool.setuptools.packages.find]
include = ["pystatkit*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
